% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceda_utils.R
\name{assign_file_type}
\alias{assign_file_type}
\title{Assign file Type}
\usage{
assign_file_type(flightFolder, lookup = NULL)
}
\arguments{
\item{flightFolder}{data.frame describing a flight folder such as that returned by \code{tidy_flight_data_check()}.}

\item{lookup}{defaults to the internal list of files described by \code{faam_file_lookup()}, but an edited 
table can be provided if the user wishes}
}
\description{
Assigns known file types from \code{file_lookup()} to files listed in a flight folder
such as that returned by \code{tidy_flight_data_check()}.
}
\author{
W. S. Drysdale
}
