% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceda_utils.R
\name{check_flight_data}
\alias{check_flight_data}
\title{Check Flight Data}
\usage{
check_flight_data(flight, verbose = TRUE)
}
\arguments{
\item{flight}{a character vector of FAAM flight numbers.}

\item{verbose}{Default true, should the available files be detailed in the console. When TRUE the data.frame is returned invisibly}
}
\description{
Performs the actual checking for \code{list_flight_data()} - Which is probably the function most will want to use when working interactively. 
This returns a nested list with some additional checks (\code{tidy_flight_data_check()}) is used to turn it into the data.frame output 
This is exported as it might have some programmatic use
}
\author{
W. S. Drysdale
}
