% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_faam_fgga.R
\name{extract_fgga_uncert}
\alias{extract_fgga_uncert}
\title{Extract fGGA Uncertainties}
\usage{
extract_fgga_uncert(filepath)
}
\arguments{
\item{filepath}{path to file}
}
\description{
The FAAM non-core fGGA instrument mena bias and uncertainty are stored in the NASA Ames header normal comments.
This function extracts the values and returns a tibble containing bias and uncertainty for CO2 and CH4.
}
\author{
W. S. Drysdale
}
