% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faam_core_summary.R
\name{faam_core_summary}
\alias{faam_core_summary}
\title{FAAM Core Summary}
\usage{
faam_core_summary(filepath)
}
\arguments{
\item{filepath}{path to file}
}
\description{
Collects metadata from the faam core and core_1hz NCDF file for browsing of 
varaibles. \cr
Collates the variable name, long_name and units with the dimention name for 
selection of the sps in \code{read_core_famm()}
}
\author{
W. S. Drysdale
}
