% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faam_file_lookup.R
\name{faam_file_lookup}
\alias{faam_file_lookup}
\title{FAAM File Lookup}
\usage{
faam_file_lookup()
}
\description{
Table of file name identifiers, and if it exists the name of the corresponding faamr function to use to read it.
}
