% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faamr-package.R
\docType{package}
\name{faamr-package}
\alias{faamr}
\alias{faamr-package}
\title{faamr: Miscellaneous functions for reading and analysing FAAM data}
\description{
Makes FAAM data analysis easier!
}
\author{
\strong{Maintainer}: Freya Squires \email{frees@bas.ac.uk} (\href{https://orcid.org/0000-0002-3364-4617}{ORCID})

Authors:
\itemize{
  \item Will Drysdale \email{will.drysdale@york.ac.uk} (\href{https://orcid.org/0000-0002-7114-7144}{ORCID})
}

}
\keyword{internal}
