% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceda_utils.R
\name{file_revision}
\alias{file_revision}
\title{File Revision}
\usage{
file_revision(name)
}
\arguments{
\item{name}{file name to extract revision from}
}
\description{
Gets the file revision number from file name.
}
\author{
W. S. Drysdale
}
