% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceda_utils.R
\name{filter_revision}
\alias{filter_revision}
\title{Filter Revision}
\usage{
filter_revision(folder)
}
\arguments{
\item{folder}{tibble representing the folder on CEDA}
}
\description{
given a flight folder data.frame, return the only the most recently revised versions of each file
}
