% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceda_utils.R
\name{flight_download}
\alias{flight_download}
\title{Flight Download}
\usage{
flight_download(flight, user, pass, dirOut, files = NULL)
}
\arguments{
\item{flight}{vector of FAAM flight numbers}

\item{user}{CEDA username}

\item{pass}{CEDA password}

\item{dirOut}{directory to save files to. Subdirectories of flight data will be made here.}

\item{files}{'fileType' names of files to be downloaded. These are shown in list_flight_data. 
If no files are supplied `flight_download()` gets all files in the flight directory and those in Core Processed and Non-core}
}
\description{
Downloads flight files from CEDA
}
\author{
W. S. Drysdale
}
