% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weight_off_wheels.R
\name{get_weight_off_wheels}
\alias{get_weight_off_wheels}
\title{Get Weight Off Wheels}
\usage{
get_weight_off_wheels(filepath)
}
\arguments{
\item{filepath}{path to file}
}
\description{
Given a path to a core ncdf file, return a dataframe of the WOW_IND (weight on wheels indicator), filtered for WOW_IND == 0 (in flight)
This is useful as WOW_IND is in the 1 Hz grid only, so the range can be supplied to \code{read_core_faam()} as \code{startDate} and \code{endDate}
}
\author{
W. S. Drysdale
}
