% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoreDateOrigin.R
\name{getCoreDateOrigin}
\alias{getCoreDateOrigin}
\title{Get Core Date Origin}
\usage{
getCoreDateOrigin(filepath)
}
\arguments{
\item{filepath}{path to file}
}
\description{
Given a path to a core ncdf file, returns the first timestamp as a nanotime
}
\author{
W. S. Drysdale
}
