% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_flight_data_local.R
\name{list_flight_data_local}
\alias{list_flight_data_local}
\title{List Flight Data Local}
\usage{
list_flight_data_local(dirContainingFlights, skipFlightListCheck)
}
\arguments{
\item{dirContainingFlights}{path to the directory that contains the flight folders. 
i.e. \code{list.dirs(dirContainingFlights)} should return a vector of valid flight numbers
These will be checked against the flight list (from \code{list_flights()}), unless \code{skipFlightListCheck}
is set to TRUE.}

\item{skipFlightListCheck}{skips checking the listed directories against \code{list_flights}. 
Useful when working without a connection to CEDA but may cause unexpeceted errors if 
directories not downloaded using \code{flight_download()} are present.}
}
\description{
Returns a tibble summarising the files that have been downloaded, which is useful for further processing
}
