% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceda_utils.R, R/list_flights.R
\name{list_flights}
\alias{list_flights}
\title{List Flights}
\usage{
list_flights(force = F)

list_flights(force = F)
}
\arguments{
\item{force}{For performance the flight list is cached in an environment variable \code{the$flightList}. Set true to force scraping CEDA.}
}
\description{
Returns a data.frame containing the flight numner, date and location on CEDA 
of all FAAM flights stored there.

Returns a data.frame containing the flight numner, date and location on CEDA 
of all FAAM flights stored there.
}
\author{
W. S. Drysdale
}
