% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalised_mean_bias.R
\name{normalised_mean_bias}
\alias{normalised_mean_bias}
\title{Calculate Normalised Mean Bias}
\usage{
normalised_mean_bias(modelled, observations, na.rm)
}
\arguments{
\item{modelled}{Column of modelled values}

\item{observations}{Column of measured values}

\item{na.rm}{a logical evaluating to TRUE or FALSE indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
NMB
}
\description{
Calculates the normalised mean bias (NMB) of modelled and measured
values according to formula in Emery 2017 (doi: 10.1080/10962247.2016.1265027)
}
\author{
Freya Squires
}
