% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalisedMeanBias.R
\name{NormalisedMeanBias}
\alias{NormalisedMeanBias}
\title{Calculate Normalised Mean Bias}
\usage{
NormalisedMeanBias(modelled, observations, na.rm)
}
\arguments{
\item{modelled}{Column of modelled values}

\item{observations}{Column of measured values}
}
\value{
NMB
}
\description{
Calculates the normalised mean bias (NMB) of modelled and measured
values according to formula in Emery 2017 (doi: 10.1080/10962247.2016.1265027)
}
\author{
Freya Squires
}
