% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceda_utils.R
\name{print_flight_data}
\alias{print_flight_data}
\title{Print flight data.}
\usage{
print_flight_data(flightDataCheck)
}
\arguments{
\item{flightDataCheck}{output of \code{check_flight_data()}}
}
\description{
Formats the output of \code{check_flight_data()} in a pretty fashion in the console.
}
\author{
W. S. Drysdale
}
