% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile25.R
\name{quantile25}
\alias{quantile25}
\title{Calculate upper and lower quantiles}
\usage{
quantile25(x, na.rm)
}
\arguments{
\item{x}{A vector}

\item{na.rm}{True or False}
}
\value{
Value
}
\description{
Calculates 25th and 75th quantiles (using percentile from stats) so it can be easily used with 'dplyr' type functions.
}
\author{
Freya Squires
}
