% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_faam_was.R
\name{read_faam_was}
\alias{read_faam_was}
\title{Read FAAM WAS files}
\usage{
read_faam_was(was_file_list)
}
\arguments{
\item{was_file_list}{A list of WAS NASA Ames file names}
}
\value{
A dataframe
}
\description{
Reads and loads WAS NASA Ames files formatted as described in:
http://cedadocs.ceda.ac.uk/73/4/index.html
}
\author{
Freya Squires
}
