% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_faam_core.R
\name{read_famm_core}
\alias{read_famm_core}
\title{Read Faam Core}
\usage{
read_famm_core(
  filepath,
  startDate = NULL,
  endDate = NULL,
  selectVar = NULL,
  sps = c(2, 4, 32, 64)[3],
  averageNanoString = NULL
)
}
\arguments{
\item{filepath}{path to file}

\item{selectVar}{vector of varibale names to load - loading all of them can take a long time!}

\item{sps}{samples per second - one of 2, 4, 32, 64. default 32}

\item{averageNanoString}{string to pass to \code{nanotime::nano_floor(date, nanotime::as.nanoduration(averageNanoString)))} for resampling date}
}
\description{
Reads either the high time resolution or 1 Hz faam core files
}
\author{
W. S. Drysdale
}
