% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flight_summary.R
\name{read_flight_summary}
\alias{read_flight_summary}
\title{Read Flight Summary}
\usage{
read_flight_summary(filepath)
}
\arguments{
\item{filepath}{path to flight summary}
}
\description{
Reads the flight summary as a tibble. Only the .txt versions are currently supported
}
\author{
W. S Drysdale
}
