% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geos_cf.R
\name{read_geos_cf}
\alias{read_geos_cf}
\title{Read GEOS Chemical Forecast (CF) model files}
\usage{
read_geos_cf(model_file_list)
}
\arguments{
\item{model_file_list}{A list of file names}
}
\value{
A dataframe
}
\description{
Reads and loads GEOS-CF model files.
}
\author{
Freya Squires
}
