% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nitrates_data.R
\name{read_nitrates_data}
\alias{read_nitrates_data}
\title{Read FAAM nitrate files}
\usage{
read_nitrates_data(nitrates_file_list, frequency = 1)
}
\arguments{
\item{nitrates_file_list}{List containing file names}

\item{frequency}{Frequency in Hz to output the data at. FAAM nirate files can be outputted at 1 or 10 Hz.}
}
\value{
Tibble
}
\description{
Reads and cleans core nitrates FAAM file.
}
\author{
Shona Wilde

Freya Squires
}
