% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_york_gc_fid_lab.R
\name{read_york_gc_fid_lab}
\alias{read_york_gc_fid_lab}
\title{Read York GC-FID Lab (WAS/SWAS)}
\usage{
read_york_gc_fid_lab(filepath, filetype = NULL)
}
\arguments{
\item{filepath}{path to file}

\item{filetype}{Is the file a .na or a .na.csv? 
If its the latter \code{repair_na_csv} to attempt to recover the data. 
By default attempts to detect from the filepath file extension}
}
\description{
Read the GC-FID runs of whole air samples
}
\author{
W. S. Drysdale
}
