% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_yorknox_data.R
\name{read_yorknox_data}
\alias{read_yorknox_data}
\title{Read York AQD NOx files}
\usage{
read_yorknox_data(nox_file_list)
}
\arguments{
\item{nox_file_list}{A list of York AQD NOx file names}
}
\value{
A dataframe
}
\description{
Reads and loads York NOx AQD files formatted as described in:
http://cedadocs.ceda.ac.uk/73/4/index.html
}
\author{
Freya Squires
}
