% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadNasaAmes.R
\name{ReadNasaAmes}
\alias{ReadNasaAmes}
\title{Read NASA Ames header}
\usage{
ReadNasaAmes(file)
}
\arguments{
\item{file}{NASA Ames file}
}
\value{
List of header length, flags, date and column names
}
\description{
Reads and extracts file information from NASA Ames headers formatted as described in:
http://cedadocs.ceda.ac.uk/73/4/index.html
Returns a list of header length, flags, file date and column names for use with other
file readers.
}
\author{
Freya Squires
}
