% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_na_csv.R
\name{repair_na_csv}
\alias{repair_na_csv}
\title{Repair .na.csv}
\usage{
repair_na_csv(filepath, outpath, NComm_header = FALSE)
}
\arguments{
\item{filepath}{path to file}

\item{outpath}{path to write repaired file to}

\item{NComm_header}{Logical default false. If the final line of normal comments contains the data header, set to true}
}
\description{
Some NASA Ames files are confused and are also .csvs. This tries to tidy them up
}
\details{
The following steps are performed
\enumerate{
\item The header length is found and lines that are greater than this have spaces replaced with "T" to fix some time stamps.
      This is changed to header length + 1 when NComm_header == TRUE
\item Lines where there are quotes (") are isolated and processed while trying to preserve 'real' commas contained within
\item commas are replaced with spaces 
\item leading and trailing white space is trimmed
\item empty rows are removed
}
}
\author{
W. S. Drysdale
}
