% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceda_utils.R
\name{tidy_flight_data_check}
\alias{tidy_flight_data_check}
\title{Tidy Flight Data Check}
\usage{
tidy_flight_data_check(flightDataCheck)
}
\arguments{
\item{flightDataCheck}{output of \code{check_flight_data()}}
}
\description{
Turns the the output of \code{check_flight_data()} into a data.frame
}
\author{
W. S. Drysdale
}
