% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceda_utils.R
\name{flight_download}
\alias{flight_download}
\title{Flight Download}
\usage{
flight_download(flight, files, dirOut, user, pass)
}
\arguments{
\item{flight}{vector of FAAM flight numbers}

\item{files}{'fileType' names of files to be downloaded. These are shown in list_flight_data}

\item{dirOut}{directory to save files to. Subdirectories of flight data will be made here.}

\item{user}{CEDA username}

\item{pass}{CEDA password}
}
\description{
Downloads flight files from CEDA
}
\author{
W. S. Drysdale
}
