% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_faam_fgga.R
\name{read_faam_fgga}
\alias{read_faam_fgga}
\title{Read FAAM fgga}
\usage{
read_faam_fgga(filepath)
}
\arguments{
\item{filepath}{path to file}
}
\description{
Reads the nasa ames files produced for the FAAM Fast Greenhouse Gas Analyser
}
\author{
W. S. Drysdale
}
